@20260807@Demons@follow新增更改计数mapper和雪花id与前端精度问题导致getprofile查询报usernotexist问题修复

+++ techdevhub-category/src/main/java/com/techdevhub/category/mapper/CategoryMapper.java
@@ -36,7 +36,7 @@ public interface CategoryMapper {
             insert into category_info (id, category_name)
             values (#{id}, #{categoryName})
             """)
-    int insert(@Param("id") Integer id, @Param("categoryName") String categoryName);
+    int insert(@Param("id") Long id, @Param("categoryName") String categoryName);
 
     @Update("""
             update category_info
